requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.35.0
decorator==4.4.2
Django==3.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
proglog==0.1.10
psycopg2-binary==2.9.8
python-dotenv==0.21.0
python-magic==0.4.27
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.0
urllib3==1.26.16
